Added a simulation of the Nd6k3 system
The user u/MisterBanzai on Reddit suggested that a simpler and
alternative to the Decibel system would be to roll a pool of
six-sided dice, keep the three largest values, and return the
sum of those three dice as the outcome of the check.  This is
a nice system and produces output distributions that are similar
to those produced by Decibel checks.

I added some code to simulate this dice mechanic so that I could
easily compare it to Decibel.